requirements.txt
aiodataloader==0.4.2
aiosmtplib==4.0.1
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
Django==4.2
django-choices-field==2.3.0
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
packaging==25.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.0
strawberry-graphql==0.273.2
strawberry-graphql-django==0.62.0
typing_extensions==4.14.0
